reorganized folder structure added simplified start job node (WIP)
